feat: refactoring ForumPage
