+++ packages/client/src/pages/Forum/components/BlockComments/BlockComment.module.scss
@@ -1,11 +1,20 @@
 @import 'src/index.module.scss';
 
+.block {
+  display: flex;
+  flex-flow: column;
+  height: 500px;
+
+  .button_publish {
+    display: flex;
+    flex-flow: column;
+    margin: 10px 50px 40px;
+  }
+}
+
 .block_topics {
   display: flex;
   flex-flow: column;
-  margin-top: 40px;
-  gap: 10px;
-  height: 400px;
 }
 
 .title {
@@ -57,3 +66,4 @@
     padding: 0px 20px;
   }
 }
+
