+++ packages/client/src/pages/Forum/components/BlockCreateTopic/BlockCreateTopic.tsx
@@ -7,12 +7,12 @@ import styles from './CreateTopic.module.scss';
 import { useNavigate } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 import { v1 } from 'uuid';
-import { ITopic } from '../Forum/components/Topic/types';
+import { ITopic } from '../../part/Topic/types';
 import { addNewTopic } from 'src/store/forum/ForumSlice';
 import { useAppSelector } from 'src/hooks/redux';
 import { dateFormatting } from 'src/utils/dateFormatting';
 
-export const ForumCreateTopic = () => {
+export const BlockCreateTopic = () => {
   const [nameTopic, setNameTopic] = useState('');
   const [descriptionTopic, setDescriptionTopic] = useState('');
   const navigate = useNavigate();
@@ -45,46 +45,30 @@ export const ForumCreateTopic = () => {
     navigate('/forum');
   }, [dispatch, login, navigate, nameTopic, descriptionTopic]);
 
-  const handlerBack = useCallback(() => {
-    navigate('/forum');
-  }, [navigate]);
-
-  const handlerLoadGame = useCallback(() => {
-    navigate('/loadinggame');
-  }, [navigate]);
-
   return (
     <ErrorBoundary>
       <div className={styles.block}>
-        <div className={styles.block_button}>
-          <div className={styles.block_buttons_top}>
-            <Button onClick={handlerBack}>GO BACK</Button>
-            <Button onClick={handlerLoadGame}>PLAY</Button>
+        <Button className={styles.button_publish} onClick={handleCreateNewTopic}>PUBLISH</Button>
+        <BlankWindow className={styles.card}>
+          <div className={styles.topic}>
+            <div className={styles.title}>NEW TOPIC: </div>
+            <Input
+              name="name_topic"
+              className={styles.input}
+              onChange={changeName}
+              value={nameTopic}
+            />
           </div>
-          <Button onClick={handleCreateNewTopic}>PUBLISH</Button>
-        </div>
-        <div className={styles.block_topics}>
-          <BlankWindow className={styles.card}>
-            <div className={styles.topic}>
-              <div className={styles.title}>NEW TOPIC: </div>
-              <Input
-                name="name_topic"
-                className={styles.input}
-                onChange={changeName}
-                value={nameTopic}
-              />
-            </div>
-            <div className={styles.title}>Topic description: </div>
-            <div className={styles.description}>
-              <Input
-                name="descrioption_topic"
-                className={styles.input}
-                onChange={changeDescription}
-                value={descriptionTopic}
-              />
-            </div>
-          </BlankWindow>
-        </div>
+          <div className={styles.title_description}>Topic description: </div>
+          <div className={styles.description}>
+            <Input
+              name="descrioption_topic"
+              className={styles.input}
+              onChange={changeDescription}
+              value={descriptionTopic}
+            />
+          </div>
+        </BlankWindow>
       </div>
     </ErrorBoundary>
   );
