+++ packages/client/src/pages/Forum/components/BlockCreateTopic/CreateTopic.module.scss
@@ -1,0 +1,54 @@
+@import 'src/index.module.scss';
+
+.block {
+  display: flex;
+  flex-flow: column;
+  height: 500px;
+}
+
+.button_publish {
+  margin: 10px 50px 40px;
+}
+
+.card {
+  padding: 8px 16px;
+  width: 100%;
+  background: $color-card;
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  display: flex;
+  box-sizing: border-box;
+  flex-flow: column;
+  height: 300px;
+}
+
+.topic {
+  display: flex;
+  flex-flow: row;
+  margin-top: 10px;
+
+  .title {
+    @include fontTitle(normal, 400, 14px, 20px);
+    color: $text-primary;
+    margin: 5px 10px;
+  }
+
+  .input {
+    height: 100%;
+  }
+}
+
+.title_description {
+  @include fontTitle(normal, 400, 14px, 20px);
+  color: $text-primary;
+  margin: 10px 10px 0px 10px;
+}
+
+.description {
+display: flex;
+flex: 1;
+
+  .input {
+    flex: 1;
+    margin: 5px 0px;
+  }
+}
