+++ packages/client/src/pages/Forum/components/BlockCreateTopic/index.ts
@@ -1,0 +1,1 @@
+export { BlockCreateTopic } from './BlockCreateTopic';
