+++ packages/client/src/pages/Forum/components/TopicList/topics.module.scss
@@ -1,0 +1,21 @@
+@import 'src/index.module.scss';
+
+.block_topics {
+  display: flex;
+  flex-flow: column;
+  height: 500px;
+
+  .button {
+    display: flex;
+    flex-flow: column;
+    margin: 10px 50px 40px;
+  }
+}
+
+.list {
+  display: flex;
+  flex-flow: column;
+  gap: 10px;
+  overflow: auto;
+  @include scroll($bg-color-main, $main_block_bg_color, $bg-color-secondary);
+}
