--- packages/client/src/pages/ForumCommentsPage/types.ts
@@ -1,5 +1,0 @@
-import { ITopic } from '../Forum/components/Topic/types';
-
-export interface Props {
-  topic: ITopic;
-}
