--- packages/client/src/pages/ForumCreateTopic/createtopic.module.scss
@@ -1,75 +1,0 @@
-@import 'src/index.module.scss';
-
-.card {
-  padding: 8px 16px;
-  width: 100%;
-  background: $color-card;
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-  display: flex;
-  box-sizing: border-box;
-  flex-flow: column;
-  height: 300px;
-}
-
-.block {
-  @include fontSecondaryText(normal, 400, 12px, 20px);
-  width: 50vw;
-  color: $text-secondary;
-  padding: 20px 0px;
-}
-
-.block_button {
-  display: flex;
-  flex-flow: column;
-  margin: 0px 50px;
-  gap: 10px;
-}
-
-.block_buttons_top {
-  display: flex;
-  flex-flow: row;
-  gap: 10px;
-  button {
-    width: 100%;
-  }
-}
-
-.block_topics {
-  display: flex;
-  flex-flow: column;
-  margin-top: 40px;
-  gap: 10px;
-  height: 400px;
-}
-
-.topic {
-  display: flex;
-  flex-flow: row;
-  margin-top: 10px;
-
-  .title {
-    @include fontTitle(normal, 400, 14px, 20px);
-    color: $text-primary;
-    margin: 5px 10px;
-  }
-
-  .input {
-    height: 100%;
-  }
-}
-
-.title {
-  @include fontTitle(normal, 400, 14px, 20px);
-  color: $text-primary;
-  margin: 10px 10px 0px 10px;
-}
-
-.description {
-  display: flex;
-  flex: 1;
-
-  .input {
-    flex: 1;
-    margin: 5px 0px;
-  }
-}
