--- packages/client/src/pages/ForumCreateTopic/index.ts
@@ -1,1 +1,0 @@
-export { ForumCreateTopic } from './ForumCreateTopic';
